avoid odd value memory access, keep data at locations divisible by 4.

+++ src/de/mindscan/furiousiron/hfb/io/BitwiseCalculations.java
@@ -33,7 +33,7 @@ import java.util.Arrays;
 public class BitwiseCalculations {
 
     // https://oeis.org/A000120 | 0x0 -- 0xf
-    private final static byte[] hexDigitWeights = { //
+    private final static int[] hexDigitWeights = { //
                     // 0, 1, 2, 3, 4, 5, 6, 7 8, 9, A, B, C, D, E, F
                     0, 1, 1, 2, // 00.[00..11]
                     1, 2, 2, 3, // 01.[00..11]
@@ -42,7 +42,7 @@ public class BitwiseCalculations {
     };
 
     // https://oeis.org/A000120 | extended to 0x00 -- 0xff 
-    private final static byte[] byteWeights = {
+    private final static int[] byteWeights = {
                     // 0x00 0x01 0x02 0x03 0x04 0x05 0x06 0x07 0x08 0x09 0x0a 0x0b 0x0c 0x0d 0x0e 0x0f
                     0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4,
                     // 0x10 0x11 0x12 0x13 0x14 0x15 0x16 0x17 0x18 0x19 0x1a 0x1b 0x1c 0x1d 0x1e 0x1f
@@ -80,14 +80,14 @@ public class BitwiseCalculations {
     /**
      * @return the byteweight
      */
-    public static byte[] getByteWeights() {
+    public static int[] getByteWeights() {
         return Arrays.copyOf( byteWeights, byteWeights.length );
     }
 
     /**
      * @return the hexdigitweight
      */
-    public static byte[] getHexDigitWeight() {
+    public static int[] getHexDigitWeight() {
         return Arrays.copyOf( hexDigitWeights, hexDigitWeights.length );
     }
